+++ src/plugins/newmail/newmail.c
@@ -107,7 +107,7 @@ gint plugin_init (gchar **error)
 		return -1;
 
 	hook_id = hooks_register_hook (MAIL_POSTFILTERING_HOOKLIST, newmail_hook, NULL);
-	if (hook_id == -1) {
+	if (hook_id == (guint) -1) {
 		*error = g_strdup (_("Failed to register newmail hook"));
 		return (-1);
 	}
